move wallet logic from cashierService into walletService.
