+++ client/src/modules/cashier/cashierService.js
@@ -2,17 +2,6 @@ import { httpClient, parseJsonResponse } from "@/lib/http";
 
 import { getPaymentByTicketId } from "../orders/ordersService";
 
-export const loginWallet = async (password) => {
-  const response = await httpClient("/wallet/auth", {
-    method: "POST",
-    headers: {
-      "Content-Type": "application/json",
-    },
-    body: JSON.stringify({ password }),
-  });
-  return await parseJsonResponse(response, null);
-};
-
 export async function getTurnOpen() {
   const response = await httpClient("/shifts");
   const shifts = await parseJsonResponse(response, []);
@@ -58,52 +47,6 @@ export async function getReport(startDate, endDate) {
   );
   return await parseJsonResponse(response, null);
 }
-
-export async function getInfo() {
-  const response = await httpClient("/wallet/getinfo");
-  return await parseJsonResponse(response, null);
-}
-
-export async function createInvoice(invoiceAmount, invoiceDesc) {
-  const response = await httpClient("/wallet/createinvoice", {
-    method: "POST",
-    headers: {
-      "Content-Type": "application/json",
-    },
-    body: JSON.stringify({
-      description: invoiceDesc,
-      amountSat: parseInt(invoiceAmount),
-    }),
-  });
-  return await parseJsonResponse(response, null);
-}
-
-export async function payInvoiceFromService(invoice) {
-  const response = await httpClient("/wallet/payinvoice", {
-    method: "POST",
-    headers: {
-      "Content-Type": "application/json",
-    },
-    body: JSON.stringify({ invoice: invoice.trim() }),
-  });
-  return await parseJsonResponse(response, null);
-}
-
-export async function getIncomingTransactions() {
-  const response = await httpClient("/wallet/payments/incoming");
-  const transactions = await parseJsonResponse(response, []);
-  return transactions ? transactions : [];
-}
-export async function getOutgoingTransactions() {
-  const response = await httpClient("/wallet/payments/outgoing");
-  const transactions = await parseJsonResponse(response, []);
-  return transactions ? transactions : [];
-}
-
-export const logoutWallet = async () => {
-  const response = await httpClient("/wallet/logout", { method: "POST" });
-  return await parseJsonResponse(response, null);
-};
 
 export async function generateReportFromData(
   startDate,
